Update Scrollers and Error Handling
(1) The MomentumScroller and SmoothScroller callbacks have been replaced
with CustomEvents that can be listened to.

(2) The MomentumScroller and SmoothScroller constructors have been
simplified, with each now only requiring a reference to the
scrollContainer. The optional parameters that were available in the
previous MomentumScroller constructor are still available but are now
accessed via chainable setters, either during or after instantiation.
Similarly, the optional parameters that were available in the previous
SmoothScroller constructor are still available but are now accessed via
the smoothScroll method or the new SmoothScroller.scroll static method
as optional parameters.

(3) SmoothScroller now allows you to scroll a scrollContainer two ways.
The previous way, which was to explicitly instantiate a SmoothScroller
and then use the smoothScroll method, is still available. This, however,
required that you use the new keyword to explicitly instantiate every
SmoothScroller. Now it is possible to scroll a scrollContainer without
explicitly instantiating one. Simply use the SmoothScroller.scroll
static method. This method automatically instantiates a SmoothScroller,
and keeps a reference to it within the static scrollerMap so that future
scrolls on the same container can be performed without having to
re-instantiate.

(4) Error handling, which used to be handled by each individual
method and function, is now handled by a function within utilities.js,
called validateArgument. validateArgument provides consistent and
flexible error handling.

(5) Fixed a bug with the MomentumScroller's edge-detection calculations;
previously, edges may not have been detected due to differences in
container and scrollbar size when overflow is set to auto or overlay.