Add New Utility and Fix validateArgument Bug
(1) getScrollableAxes was added to ScrollContainerTools

(2) The allowedValues parameter within validateArgument() now defaults
to null instead of an empty array. The old logic assumed that an empty
array meant "do not test this". However, that was illogical! An empty
array means there are no allowed values, which may be exactly what one
wants. This bug was fixed by making null the default value for
allowedValues. Now, allowedValues is not tested if it is null, but it is
tested if it is an empty array.