Update getScrollableAxes
(1) Overflow properties are now considered before making a determination
on whether an axis is scrollable or not. To be considered scrollable, an
element must have overflow value of "auto", "scroll", or "overlay".
Elements with overflow values of "hidden" are, by default, not
considered scrollable, but this can be changed by setting the parameter
ignoreElementsWithOverflowHidden to false.

(2) An exception to the overflow criteria is made for the document root
element (html element) since it is scrollable even with an overflow
value of "visible".