Remove Unnecessary Events and Fix Event Properties
(1) Removed "momentumScrollerPointerDown" and
"momentumScrollerPointerUp" since these were basically rebranded
"pointerdown" and "pointerup" events, which are events that anyone can
easily listen for if they need to.

(2) Properties added to the CustomEvent details object have been revised
and most now use NaN as a default value instead of null; null was not
a good choice in this case because JavaScript converts it to zero, which
was causing some miscalculations.

(3) autoCreateScrollers and createScroller static method's parameter
called "autoActivate" was renamed "activateImmediately" for clarity; it
was also moved to an options object.

(4) "keepCurrentSelectors" was moved to an options object in the setters
that use it as a parameter.