Update ScrollContainerTools
(1) The getEdgeStatus static method now checks page progression to
properly define left and right edges. Previously, the function defined
left and right edges with an assumption that an element was
left-to-right. Now it does not make that assumption. Additionally, it
includes an options object parameter so that a cachedPageProgression
may be included. This is useful in cases where getEdgeStatus is called
frequently because it skips having to call getComputedStyle to calculate
page progression.

(2) The getPageProgression static method was added. It determines if an
element is "left-to-right" or "right-to-left" based on the following
specification: https://www.w3.org/TR/css-writing-modes-4/#page-direction

(3) The getScrollableAxes method was renamed getAxisOverflowProperties.
It makes fewer assumptions on scrollability and instead provides a few
more data points to help the caller make more informed choices about
scrollability.

(4) containElement was added. It is useful for applying containers, i.e.
wrappers.