Improve Scroller Compatibility and More
(1) The autoCreateScrollers static method has several changes. The
coerceBodyToScroller parameter was removed in favor or better processing
of the root/body elements. A new parameter, selectorsToIgnore, was
added to give users a way to prevent processing of certain selectors by
the method. The method no longer modifies the overflow, height, width,
or margin properties of the root or body elements.

(2) A "counter-bounce" mechanism was added. It is necessary because
translation of the <html> element past its block-start or inline-start
causes clipping, which reduces scrollHeight/scrollWidth, and thus does
not cause a bouncing effect. This mechanism counters the reduction in
scrollHeight/scrollWidth caused by the clipping, and thereby allows the
bouncing effect to occur.
Relevant specification:
https://www.w3.org/TR/css-overflow-3/#scrolling-direction

(3) The thresholdTest default threshold was reduced from 10 pixels to 5
pixels. This was done because fast flicks would sometimes not be
detected, causing scrolling to stop unexpectedly. A 5 pixel threshold
should be enough wiggle room for the elements that need it.

(4) momentumScrollerRoute was renamed momentumScrollerPointerRoute. It
returns an object with 3 properties: pointerEvent, routeTo, routeFrom.

(5) Page progression is now checked during initialization and cached
so that the cached value can be provided to the getEdgeStatus method
of ScrollContainerTools.

(6) The "momentumScrollerPointerRoute" listener is now more readable. It
also tells the "routeFrom" scroller to delay cursor changes until after
a "momentumScrollerPointerHandlingStop" event. This fixes a bug where
the cursor would sometimes flicker between the "grab" and "grabbing"
states as one momentum scroller was handing off the pointer to another
momentum scroller.

(7) setAllowHorizontalScrolling and setAllowVerticalScrolling no longer
modify overflow properties.

(8) #updateBouncePosition private instance method was added

(9) Private property #bounceFallingOnlyX was renamed #bounceReboundOnlyX
and #bounceFallingOnlyY was renamed #bounceReboundOnlyY