Change #pointerDownRouter to _pointerDownRouter
(1) The pointerdown event listener that calls this.#pointerDownRouter
irregularly throws an uncaught TypeError: "Receiver must be class %",
where % is an instance's private property,
"at HTMLDocument.<anonymous>". This seems to only happen in
Chromium-based browsers.
I have yet to find a solution to this, so for now #pointerDownRouter
will become _pointerDownRouter, which seems to resolve the issue in
Chromium-based browsers.