Add Click preventDefault for Gecko, More
(1) When a pointerdown event starts on an anchor and then moves outside
of the threshold, there should no longer be a click event. Chromium
browsers do not dispatch click events, but Gecko browsers do. Therefore,
a fix was implemented to make sure Gecko clicks that have moved outside
of the threshold have preventDefault called.

(2) Added static #activeScrollers and #inactiveScrollers sets to cache
these values instead of having to query all scrollers.

(3) Also added static #scrollerHandlingPointer for the same reason.

(4) Made static verifySelectors a private method.

(5) Adjusted deceleration and bounciness levels.

(6) Removed "blur" event listeners from individual instances.

(7) Adjusted the overscroll calculation so that when the user contracts
the overscroll effect, the pointermove:translate ratio is 1:1 instead
of an exponential decay. The exponential decay still applies when the
overscroll is expanding.