Allow Touch-Screen Devices
(1) Previously, a full-screen modal element would display if a
touch-screen device was detected, with a message asking the user to
return on a non-touch-screen device with a mouse. This is no longer
how MomentaMouse works. Now, as long as a mouse is detected, it works.
It even works on touch-screens if a mouse is available. The page now
uses CSS media queries as heuristics to identify if a fine, hoverable
pointer is available. Although this is not mouse-specific (it may also
accept touchpads), it is the best solution for now. When a mouse is
found, MomentaMouse automatically activates. When a mouse is not found,
MomentaMouse disables itself and the page updates to reflect this.