Make Page Responsive To Mouse Availability
(1) MomentaMouse now works as long as a mouse is detected. It even works
on touch-screens if a mouse is available. The page now uses CSS media
queries as heuristics to identify if a fine, hoverable pointer is
available. Although this is not mouse-specific (it may also accept
touchpads), it is the best solution for now. When a mouse is found,
MomentaMouse automatically activates. When a mouse is not found,
MomentaMouse disables itself and the page updates to reflect this.