Update MomentaMouse Argument
(1) The Charm's Jobs section includes a video gallery with three videos.
Each video has a progress bar associated with it above. Although they
were clickable, they didn't do anything, and this was because they had
not been included in one of the MomentaMouse parameters. This behavior
has been corrected. Now, when a progress bar is clicked, the gallery
scrolls to the desired video.