Refactor range Function
(1) New function signature. It now takes a required "stop" argument and
optional object argument for specifying start and step. JavaScript does
not support overloaded functions, so, although the previous
implementation mimicked Python's overloaded analog, it was not ideal
because if you were to supply a single argument, it would still be
assigned to "start" even though it should be assigned to "stop." Now
there is no longer a need for solutions like variable reassignment or
recursive calling with corrected arguments.

(2) Made the range start-inclusive and end-exclusive so that the return
matches what Python's range function would return. Based this decision
on Dijkstra's analysis of zero-based indexing:
https://www.cs.utexas.edu/users/EWD/transcriptions/EWD08xx/EWD831.html

(3) Negative step arguments are now possible for producing descending
ranges.